+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Allow specifying an explicit script type for xpubs using `-x <xpub>:wpkh`, `-x <xpub>:shwpkh` or `-x <xpub>:pkh`.
+
 ## 0.2.3 - 2021-03-17
 
 - New `--prune-until <target>` option to automatically prune the chain up to the given target (height, unix timestamp or YYYY-MM-DD formatted date)